requirements.txt
python-slugify
psycopg2
